+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - `c.expectation_ps()` now also supports `ps` argument directly (pauli structures)
 
+- Add tc version print in `tc.about()` method
+
 ### Fixed
 
 - `tc.results.counts.plot_histogram` now can dispatch kws to corresponding qiskit method